Flechas ya se guardan
solo falta el bug que una vez has guardado una partida cuando quieres salir de pause te obliga a guardar

+++ HolaSDL/SDLGameObject.h
@@ -18,7 +18,7 @@ protected:
 public:
 
 	virtual void saveToFile(ofstream* output) {
-		*output << posicion.GetX() << " " << posicion.GetY() << "\n";
+		*output << (int)posicion.GetX() << " " << (int)posicion.GetY() << "\n";
 	};
 	virtual void loadFromFile(ifstream* input) {
 		int x, y;
